feat: change node-sass to dart-sass, fix typescript error
Node sass was depreciated as of last year and always casued limits on the node version, since it requried a node version but angular required the lower version because of the sass builder. Changing to dart-sass fixes all of this and is the main repo as of last year.

Fixed typescript error as it should be locked to the excact version we use not the point version